Added new command to clone nextjs waas boilerplate
New command: '-CliBaseCommand- boilerplates create-kit-embedded-wallet-nextjs-starter'
